♻️ refactor(tests): fold per-issue test files into topical ones
The per-issue integration files (issue_347, _384, _572, _599 and the
type-aliases copy) each duplicated the @expected decorator and runner
from test_integration.py and carried their own Sphinx roots whose conf
matched test-integration's. Five files and three roots for tests that
all have a topical home.

Move the rendering tests into test_integration.py's harness as OPTIONS
on each function, the documented-type-alias and forward-ref builds into
test_pep695.py, the format_annotation units into test_annotations.py,
the resolver fallbacks into test_resolver/test_type_hints.py, and the
numpydoc-extension builds into test_formats_numpydoc.py as one
parametrized test. The forward-ref and eager modules now live in the
test-integration root; the numpydoc root keeps its distinct conf under
a topical name. No test semantics changed except expected output where
the shared harness forces always_use_bars_union.
