+++ pyproject.toml
@@ -122,11 +122,11 @@ lint.per-file-ignores."tests/**/*.py" = [
   "UP007",   # we test for old Union syntax
   "UP045",   # we test for old Optional syntax
 ]
-lint.per-file-ignores."tests/roots/test-issue_572/mod_forward_ref.py" = [
+lint.per-file-ignores."tests/roots/test-integration/mod_forward_ref.py" = [
   "I002",    # intentionally omits `from __future__ import annotations` to test forward references
   "PLR6301", # method must be instance method to test class forward refs
 ]
-lint.per-file-ignores."tests/roots/test-issue_599/mod_eager.py" = [
+lint.per-file-ignores."tests/roots/test-integration/mod_eager.py" = [
   "I002", # this file intentionally omits `from __future__ import annotations` to test non-deferred annotations
 ]
 lint.per-file-ignores."tests/roots/test-pyi-stubs/stub_mod.py" = [
