--- tests/roots/test-issue_572/conf.py
@@ -1,14 +1,0 @@
-from __future__ import annotations
-
-import pathlib
-import sys
-
-sys.path.insert(0, str(pathlib.Path(__file__).parent))
-
-master_doc = "index"
-
-extensions = [
-    "sphinx.ext.autodoc",
-    "sphinx.ext.napoleon",
-    "sphinx_autodoc_typehints",
-]
