--- tests/roots/test-issue_572/index.rst
@@ -1,2 +1,0 @@
-Test
-====
