--- tests/roots/test-issue_599/index.rst
@@ -1,2 +1,0 @@
-Test
-====
