+++ tests/test_resolver/test_type_hints.py
@@ -12,6 +12,9 @@ from pathlib import Path
 from typing import Any, Union, get_args, get_origin
 from unittest.mock import MagicMock, patch
 
+if sys.version_info >= (3, 14):  # pragma: >=3.14 cover
+    import annotationlib
+
 import pytest
 
 from sphinx_autodoc_typehints._annotations import MyTypeAliasForwardRef
@@ -221,3 +224,38 @@ def test_stub_annotations_not_polluted_on_repeated_calls(tmp_path: Path) -> None
         for name in [n for n in sys.modules if n.startswith("stubpkg")]:
             del sys.modules[name]
         _TYPE_GUARD_IMPORTS_RESOLVED.discard("stubpkg.mod")
+
+
+@pytest.mark.skipif(sys.version_info < (3, 14), reason="annotationlib requires Python 3.14+")
+def test_get_type_hint_uses_annotationlib_on_name_error() -> None:  # pragma: >=3.14 cover
+    """_get_type_hint falls back to annotationlib.get_annotations on 3.14+ NameError."""
+    sentinel = {"x": int}
+
+    def dummy() -> None: ...
+
+    with (
+        patch(
+            "sphinx_autodoc_typehints._resolver._type_hints.get_type_hints",
+            side_effect=NameError("name 'Foo' is not defined"),
+        ),
+        patch.object(annotationlib, "get_annotations", return_value=sentinel) as mock_get_ann,
+    ):
+        result = _get_type_hint([], "dummy", dummy, {})
+
+    mock_get_ann.assert_called_once_with(dummy, format=annotationlib.Format.FORWARDREF)
+    assert result is sentinel
+
+
+@pytest.mark.skipif(sys.version_info >= (3, 14), reason="Tests pre-3.14 fallback path")
+def test_get_type_hint_falls_back_to_dunder_annotations_before_314() -> None:  # pragma: <3.14 cover
+    """_get_type_hint falls back to __annotations__ on pre-3.14 NameError."""
+
+    def dummy(x: int) -> str: ...
+
+    with patch(
+        "sphinx_autodoc_typehints._resolver._type_hints.get_type_hints",
+        side_effect=NameError("name 'Foo' is not defined"),
+    ):
+        result = _get_type_hint([], "dummy", dummy, {})
+
+    assert result == dummy.__annotations__
